refactor: Reorganize database structure following clean architecture
- Move database related code to data.database package
- Organize entities in dedicated package
- Implement clean entity structure
- Add proper type converters
- Improve code organization and maintainability
- Create clear table structure with proper relationships
