+++ app/src/main/java/com/github/cstrerath/uncover/ExperienceManager.kt
@@ -2,6 +2,7 @@ package com.github.cstrerath.uncover
 
 import android.content.Context
 import android.util.Log
+import com.github.cstrerath.uncover.data.database.AppDatabase
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
