+++ app/src/main/java/com/github/cstrerath/uncover/data/database/dao/LocationDao.kt
@@ -1,10 +1,11 @@
-package com.github.cstrerath.uncover
+package com.github.cstrerath.uncover.data.database.dao
 
 import androidx.room.Dao
 import androidx.room.Delete
 import androidx.room.Insert
 import androidx.room.Query
 import androidx.room.Update
+import com.github.cstrerath.uncover.data.database.entities.Location
 
 @Dao
 interface LocationDao {
