+++ app/src/main/java/com/github/cstrerath/uncover/data/database/entities/QuestStep.kt
@@ -1,4 +1,4 @@
-package com.github.cstrerath.uncover
+package com.github.cstrerath.uncover.data.database.entities
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
