+++ app/src/main/java/com/github/cstrerath/uncover/domain/managers/QuestManager.kt
@@ -3,7 +3,7 @@ package com.github.cstrerath.uncover.domain.managers
 
 import android.content.Context
 import android.util.Log
-import com.github.cstrerath.uncover.AppDatabase
+import com.github.cstrerath.uncover.data.database.AppDatabase
 import com.github.cstrerath.uncover.QuestProgressManager
 
 class QuestManager(context: Context) {
