+++ app/src/main/java/com/github/cstrerath/uncover/ui/activities/QuestActivity.kt
@@ -22,19 +22,19 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.github.cstrerath.uncover.AppDatabase
-import com.github.cstrerath.uncover.CharacterClass
-import com.github.cstrerath.uncover.CharacterQuestProgress
-import com.github.cstrerath.uncover.CharacterQuestProgressDao
-import com.github.cstrerath.uncover.GameCharacterDao
-import com.github.cstrerath.uncover.LocationDao
-import com.github.cstrerath.uncover.QuestDao
+import com.github.cstrerath.uncover.data.database.AppDatabase
+import com.github.cstrerath.uncover.data.database.entities.CharacterClass
+import com.github.cstrerath.uncover.data.database.entities.CharacterQuestProgress
+import com.github.cstrerath.uncover.data.database.dao.CharacterQuestProgressDao
+import com.github.cstrerath.uncover.data.database.dao.GameCharacterDao
+import com.github.cstrerath.uncover.data.database.dao.LocationDao
+import com.github.cstrerath.uncover.data.database.dao.QuestDao
 import com.github.cstrerath.uncover.QuestProgressManager
 import com.github.cstrerath.uncover.QuestResources
-import com.github.cstrerath.uncover.QuestStage
+import com.github.cstrerath.uncover.data.database.entities.QuestStage
 import com.github.cstrerath.uncover.data.database.dao.QuestStepDao
 import com.github.cstrerath.uncover.R
-import com.github.cstrerath.uncover.StepType
+import com.github.cstrerath.uncover.data.database.entities.StepType
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
